Add collaboration-move-stock-waste
* Explanation: The user asks the view to "move" the top 3 cards from the
stak to the waste, as in klondike you can only move "stacks" (of 1 or 3
cards) from the stock to the waste. Then the view asks to the controller
to do the action, which will be reflected on the model. Then to update
the interface, the view will ask the model about the stock and the
waste.